ADD: refactor
