+++ src/main/scala/trace/TraceSinkAlways.scala
@@ -20,7 +20,8 @@ class WithTraceSinkAlways(targetId: Int = 0) extends Config((site, here, up) => 
   case TilesLocated(InSubsystem) => up(TilesLocated(InSubsystem), site) map {
     case tp: RocketTileAttachParams => {
       tp.copy(tileParams = tp.tileParams.copy(
-        ltrace = Some(tp.tileParams.ltrace.get.copy(buildSinks = tp.tileParams.ltrace.get.buildSinks :+ (p => (LazyModule(new TraceSinkAlways()(p)), targetId)))))
+        traceParams = Some(tp.tileParams.traceParams.get.copy(buildSinks = 
+          tp.tileParams.traceParams.get.buildSinks :+ (p => (LazyModule(new TraceSinkAlways()(p)), targetId)))))
       )
     }
     case other => other
